Finishing 6th homework (#8)
* Finishing 6th homework

* changing main

+++ src/main/java/org/example/main/controller/ItemController.java
@@ -30,13 +30,13 @@ public class ItemController {
         itemService.addItem(itemDto);
     }
     @DeleteMapping("/delete/{id}")
-    public void deleteById(@PathVariable int id) {
+    public void deleteById(@PathVariable Long id) {
         itemService.delete(id);
 
     }
 
     @PostMapping("/{id}/edit")
-    public void editUpdate(@PathVariable(value = "id") int id, String jsonString) {
+    public void editUpdate(@PathVariable(value = "id") Long id, String jsonString) {
         ItemDto itemDto = jsonMapper.convertFromJsonString(jsonString, ItemDto.class);
         itemService.update(id, itemDto);
     }
