+++ src/main/java/org/example/main/controller/MatchController.java
@@ -30,13 +30,13 @@ public class MatchController {
         matchService.addMatch(matchDto);
     }
     @DeleteMapping("/delete/{id}")
-    public void deleteById(@PathVariable int id) {
+    public void deleteById(@PathVariable Long id) {
         matchService.delete(id);
 
     }
 
     @PostMapping("/{id}/edit")
-    public void editUpdate(@PathVariable(value = "id") int id, String jsonString) {
+    public void editUpdate(@PathVariable(value = "id") Long id, String jsonString) {
         MatchDto matchDto = jsonMapper.convertFromJsonString(jsonString, MatchDto.class);
         matchService.update(id, matchDto);
     }
